requirements.txt
flask
discord.py
pandas
numpy
matplotlib
seaborn


